Translation unit for func "min"
