+++ chistyakov.vladislav/P4/min.hpp
@@ -1,0 +1,7 @@
+#ifndef MIN_HPP
+#define MIN_HPP
+
+int min(int num1, int num2, int num3, int num4);
+int min(int num1, int num2, int num3);
+
+#endif
